works but why

+++ src/main/java/com/j9nos/RAM.java
@@ -7,13 +7,13 @@ import java.lang.management.ManagementFactory;
 public final class RAM {
     private RAM() {
     }
-    
-    public static long usage(){
+
+    public static int usage() {
         final OperatingSystemMXBean os = (OperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
         final long total = os.getTotalMemorySize();
         final long free = os.getFreeMemorySize();
         final long used = total - free;
-        return Math.round((double) used / total * 100);
+        return (int) Math.round((double) used / total * 100);
     }
 
 }
